Time: 15 ms (97.29%), Space: 82.6 MB (55.13%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.java
@@ -1,6 +1,6 @@
 class Solution {
     public int minDays(int[] bloomDay, int m, int k) {
-        int val= m*k;
+        long val= (long)m*k;
         if(val> bloomDay.length)return -1;
         int start=min(bloomDay);
         int end= max(bloomDay);
